Thesis
This is the thesis for the project.
It comprises of the literature review, introduction, methodology used and the implementation details of the project. It also includes the results and discussions from the implementation and experimentation phases. Finally it provides the conclusion and future scope of the subject followed by the references.